+++ v3/crates/metadata-resolve/src/stages/commands/command.rs
@@ -57,7 +57,7 @@ pub fn resolve_command(
                 get_argument_kind(&argument.argument_type, subgraph, boolean_expression_types);
             if arguments
                 .insert(
-                    argument.name.clone(),
+                    argument.name.value.clone(),
                     ArgumentInfo {
                         argument_type: mk_qualified_type_reference(
                             &argument.argument_type,
@@ -65,6 +65,7 @@ pub fn resolve_command(
                         ),
                         argument_kind,
                         description: argument.description.clone(),
+                        path: argument.name.path.clone(),
                     },
                 )
                 .is_some()